requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-allauth==0.41.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-formtools==2.2
django-import-export==2.0.2
django-js-asset==1.2.2
django-model-utils==4.0.0
django-mptt==0.11.0
django-notifications-hq==1.6.0
django-password-reset==2.0
django-reversion==3.0.5
django-simpleui==3.9
django-taggit==1.2.0
future==0.18.2
html5lib==1.0.1
httplib2==0.9.2
idna==2.9
jsonfield==3.1.0
lxml==4.4.2
Markdown==3.2.1
oauthlib==3.1.0
Pillow==7.0.0
Pygments==2.6.1
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.1
swapper==1.1.2.post1
tablib==1.1.0
urllib3==1.25.8
webencodings==0.5.1
xadmin==0.6.1
